+++ KtistaClient/app/src/main/java/com/lkorasik/ktistaclient/MainActivity.kt
@@ -23,14 +23,14 @@ class MainActivity : AppCompatActivity() {
         val navView: BottomNavigationView = binding.navView
 
         val navController = findNavController(R.id.nav_host_fragment_activity_main)
-        // Passing each menu ID as a set of Ids because each
-        // menu should be considered as top level destinations.
+
         val appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.navigation_feed, R.id.navigation_dashboard, R.id.navigation_notifications
+                R.id.navigation_feed, R.id.navigation_profile, R.id.navigation_settings
             )
         )
-        setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
+        setupActionBarWithNavController(navController, appBarConfiguration)
+
     }
 }
