+++ KtistaClient/app/src/main/java/com/lkorasik/ktistaclient/ui/profile/ProfileFragment.kt
@@ -1,0 +1,37 @@
+package com.lkorasik.ktistaclient.ui.profile
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.lkorasik.ktistaclient.databinding.FragmentProfileBinding
+
+class ProfileFragment : Fragment() {
+
+    private lateinit var profileViewModel: ProfileViewModel
+    private var _binding: FragmentProfileBinding? = null
+
+    private val binding get() = _binding!!
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        profileViewModel =
+            ViewModelProvider(this)[ProfileViewModel::class.java]
+
+        _binding = FragmentProfileBinding.inflate(inflater, container, false)
+
+        /*
+        val textView: TextView = binding.textDashboard
+        dashboardViewModel.text.observe(viewLifecycleOwner, Observer {
+            textView.text = it
+        })
+         */
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
