+++ KtistaClient/app/src/main/java/com/lkorasik/ktistaclient/ui/profile/ProfileViewModel.kt
@@ -1,10 +1,10 @@
-package com.lkorasik.ktistaclient.ui.dashboard
+package com.lkorasik.ktistaclient.ui.profile
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class ProfileViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is dashboard Fragment"
