+++ KtistaClient/app/src/main/java/com/lkorasik/ktistaclient/ui/settings/SettingsFragment.kt
@@ -1,0 +1,31 @@
+package com.lkorasik.ktistaclient.ui.settings
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.lkorasik.ktistaclient.databinding.FragmentSettingsBinding
+
+class SettingsFragment : Fragment() {
+
+    private lateinit var settingsViewModel: SettingsViewModel
+    private var _binding: FragmentSettingsBinding? = null
+
+    private val binding get() = _binding!!
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        settingsViewModel =
+            ViewModelProvider(this)[SettingsViewModel::class.java]
+
+        _binding = FragmentSettingsBinding.inflate(inflater, container, false)
+
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
