+++ KtistaClient/app/src/main/java/com/lkorasik/ktistaclient/ui/settings/SettingsViewModel.kt
@@ -1,10 +1,10 @@
-package com.lkorasik.ktistaclient.ui.notifications
+package com.lkorasik.ktistaclient.ui.settings
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class SettingsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is notifications Fragment"
