+++ KtistaClient/app/src/main/res/menu/bottom_nav_menu.xml
@@ -2,18 +2,18 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/navigation_feed"
+        android:id="@+id/feed"
         android:icon="@drawable/ic_home_black_24dp"
         android:title="@string/title_feed" />
 
     <item
-        android:id="@+id/navigation_dashboard"
-        android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard" />
+        android:id="@+id/profile"
+        android:icon="@drawable/ic_notifications_black_24dp"
+        android:title="@string/title_profile" />
 
     <item
-        android:id="@+id/navigation_notifications"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications" />
+        android:id="@+id/settings"
+        android:icon="@drawable/ic_dashboard_black_24dp"
+        android:title="@string/title_settings" />
 
 </menu>
