+++ KtistaClient/app/src/main/res/values/strings.xml
@@ -1,8 +1,8 @@
 <resources>
     <string name="app_name">Ktista</string>
     <string name="title_feed">Feed</string>
-    <string name="title_dashboard">Dashboard</string>
-    <string name="title_notifications">Notifications</string>
+    <string name="title_settings">Settings</string>
+    <string name="title_profile">Profile</string>
 
     <string name="settings_first_name">First name</string>
     <string name="settings_last_name">Last name</string>
